Use long for deltas in model classes

+++ src/main/java/eu/eutampieri/catacombs/model/GameObject.java
@@ -34,7 +34,7 @@ public abstract class GameObject {
      * 
      * @param delta time between updates
      */
-    public abstract void update(float delta);
+    public abstract void update(long delta);
 
     /**
      * Renders object with the corresponding sprite.
